+++ src/App.vue
@@ -13,10 +13,11 @@
 <script setup>
 import { ref } from "vue";
 
-const selectedPreview = ref("button-group-preview");
+const selectedPreview = ref("donut-chart-preview");
 
 const previewOptions = [
 	{ label: "button-group", value: "button-group-preview" },
+	{ label: "donut-chart", value: "donut-chart-preview" },
 	{ label: "form", value: "form-preview" },
 	{ label: "form-input", value: "form-input-preview" },
 	{ label: "link-tag", value: "link-tag-preview" },