+++ src/components/form/form-actions/form-actions.vue
@@ -3,3 +3,10 @@
 		<slot />
 	</div>
 </template>
+
+<script>
+/**
+ * `form-actions` is a functional component that adds consistent layout of the
+ * actions of a form.
+ */
+</script>