+++ src/components/form/form-layout/form-layout.vue
@@ -3,3 +3,10 @@
 		<slot />
 	</div>
 </template>
+
+<script>
+/**
+ * `form-layout` is a functional component that adds consistent spacing between
+ * elements in a form.
+ */
+</script>